+++ R/00_warpgeneral.R
@@ -93,7 +93,7 @@ gdal_raster_data <- function(dsn, target_crs = NULL, target_dim = NULL, target_e
                              silent = FALSE, band_output_type = band_output_type, 
                              options = options, 
                              dsn_outname = "", 
-                    include_meta = include_meta)
+                    include_meta = include_meta, nara = FALSE)
 }
 
 #' @name gdal_raster_data
@@ -153,7 +153,7 @@ gdal_raster_dsn <- function(dsn, target_crs = NULL, target_dim = NULL, target_ex
                             resample = resample, 
                             silent = FALSE, band_output_type = band_output_type, 
                             options = options, 
-                            dsn_outname = out_dsn[1L], include_meta = include_meta)
+                            dsn_outname = out_dsn[1L], include_meta = include_meta, nara = FALSE)
 }
 
 #' @name gdal_raster_data
@@ -166,7 +166,7 @@ gdal_raster_image <- function(dsn, target_crs = NULL, target_dim = NULL, target_
   if (is.null(target_ext)) target_ext <-  numeric()
   if (is.null(target_dim)) target_dim <- integer() #info$dimension
   if (is.null(target_res)) target_res <- numeric() ## TODO
-  if (is.null(band_output_type)) band_output_type <- "UInt8"
+  if (is.null(band_output_type)) band_output_type <- "Byte"
   if (is.null(bands)) {
     nbands <- vapour_raster_info(dsn[1])$bands
     bands <- seq(min(c(nbands, 4L)))
@@ -179,10 +179,38 @@ gdal_raster_image <- function(dsn, target_crs = NULL, target_dim = NULL, target_
                             bands = bands, 
                             resample = resample, 
                             silent = FALSE, band_output_type = band_output_type, 
-                            options = options, dsn_outname = "", include_meta = include_meta)
+                            options = options, dsn_outname = "", include_meta = include_meta, nara = FALSE)
   atts <- attributes(bytes)
   out <- list(as.vector(grDevices::as.raster(array(unlist(bytes, use.names = FALSE), c(length(bytes[[1]]), 1, max(c(3, length(bytes))))))))
   attributes(out) <- atts
+  out
+}
+
+
+#' @name gdal_raster_data
+#' @export
+gdal_raster_nara <- function(dsn, target_crs = NULL, target_dim = NULL, target_ext = NULL, target_res = NULL, 
+                              resample = "near", bands = NULL, band_output_type = NULL, options = character(), include_meta = TRUE) {
+  
+  if (length(target_res) > 0 ) target_res <- as.numeric(rep(target_res, length.out = 2L))
+  if (is.null(target_crs)) target_crs <- "" 
+  if (is.null(target_ext)) target_ext <-  numeric()
+  if (is.null(target_dim)) target_dim <- integer() #info$dimension
+  if (is.null(target_res)) target_res <- numeric() ## TODO
+  band_output_type <- "Byte"
+  if (is.null(bands)) {
+    nbands <- vapour_raster_info(dsn[1])$bands
+    bands <- seq(min(c(nbands, 4L)))
+  }
+  include_meta <- isTRUE(include_meta)
+  out <- warp_general_cpp(dsn, target_crs, 
+                            target_ext, 
+                            target_dim, 
+                            target_res, 
+                            bands = bands, 
+                            resample = resample, 
+                            silent = FALSE, band_output_type = band_output_type, 
+                            options = options, dsn_outname = "", include_meta = include_meta, nara = TRUE)
   out
 }