+++ inst/include/gdalreadwrite/gdalreadwrite.h
@@ -204,7 +204,7 @@ inline CharacterVector gdal_create_copy(CharacterVector dsource, CharacterVector
 }
 inline List gdal_read_block(CharacterVector dsn, IntegerVector offset,
                             IntegerVector dimension, IntegerVector band,
-                            CharacterVector band_output_type, LogicalVector unscale) {
+                            CharacterVector band_output_type, LogicalVector unscale, LogicalVector nara) {
   IntegerVector window(6);
   window[0] = offset[0];
   window[1] = offset[1];
@@ -212,7 +212,7 @@ inline List gdal_read_block(CharacterVector dsn, IntegerVector offset,
   window[3] = dimension[1];
   window[4] = dimension[0];
   window[5] = dimension[1];
-  return   gdalraster::gdal_raster_io(dsn, window, band, "nearestneighbour", band_output_type, unscale);
+  return   gdalraster::gdal_raster_io(dsn, window, band, "nearestneighbour", band_output_type, unscale, nara);
 }
 
 inline LogicalVector gdal_write_block(CharacterVector dsn, NumericVector data,